Added wall textures, but they don't look good, are a bit broken, and lag the game.

+++ core/src/uk/co/samatkins/gbjam4/Entity.java
@@ -2,12 +2,12 @@ package uk.co.samatkins.gbjam4;
 
 public class Entity implements Comparable<Entity> {
 
-	public final int x, y;
+	public final float x, y;
 	public final GBImage image;
 
 	public float distanceToPlayer;
 
-	public Entity(int x, int y, GBImage image) {
+	public Entity(float x, float y, GBImage image) {
 		this.x = x;
 		this.y = y;
 		this.image = image;
